firmware: tegra: simplify module loading
Currently, module loading requires 4 writes to different debugfs attrs.

	cat /sys/kernel/debug/bpmp/module/load
	echo 1 > /sys/class/firmware/bpmp.mod.<id>/loading
	cat <module-path> > /sys/class/firmware/bpmp.mod.<id>/data
	echo 0 > /sys/class/firmware/bpmp.mod.<id>/loading

This is tedious. Also the modules are not named under the
/sys/kernel/debug/bpmp/module/ tree, but instead they are enumerated as
bpmp.mod.1, bpmp.mod.2..etc., making it harder to know which is which.
Additionally, the request_firmware_nowait() when used with uevent =
false, is causing system suspend to fail.

Therefore, simplify the module loading sematics slightly, by requiring
only to echo the module name into /sys/kernel/debug/bpmp/module/load.
The kernel automatically picks the module from one of its pre-defined
paths. This makes it easy for someone wanting to do a quick module load
and test.

The new sematics are:
	echo <module-name> > /sys/kernel/debug/bpmp/module/load # load
	cat /sys/kernel/debug/bpmp/module/<module-name>/unload  # unload

This simplification comes with small sacrifices, none of which are
important in Tegra platforms.

- module binary must be present in one of the pre-defined locations
  (defined in drivers/base/firmware_class.c)
        "/system/etc/firmware",
        "/system/vendor/etc/firmware",
        "/lib/firmware/updates/" UTS_RELEASE,
        "/lib/firmware/updates",
        "/lib/firmware/" UTS_RELEASE,
        "/lib/firmware"
  (This can be worked around by changing the fw_path_para module
  parameter).
- A module can be loaded only once
- Multiple modules with the same name is not supported

Change-Id: I7c65af070f018398cf719e3f72fb85d2044eac0a
Signed-off-by: Sivaram Nair <sivaramn@nvidia.com>