pm: l4t: honor events_check_enabled for wakeup
An Android-specific hack was introduced to work
with Android wake_locks in this commit:
9f1ab7af88fe "unknown changes from android-tegra-nv-3.4"

This hack effectively ignores events_check_enabled flag for wakeup
events, and always counts pending wakeup events. This is blocking
suspend for some devices on Linux. The hack should be enclosed in
CONFIG_ANDROID as indicated by the comment.

Bug 200114050

Change-Id: I845cae13107fe744c9d886ff4a691cbaab02c199
Signed-off-by: Zheng Liu <zhliu@nvidia.com>
Reviewed-on: http://git-master/r/775971
Reviewed-by: Stephen Warren <swarren@nvidia.com>
Reviewed-by: Matthew Pedro <mapedro@nvidia.com>
Reviewed-by: Winnie Hsu <whsu@nvidia.com>